UI consistency and assertion fixes.
* Use "Close" in the file menu of every main frame window instead of the
  mismatch of the terms "Close", "Exit", and "Quit".
* Change the terminology in the Pcbnew graphic text edit dialog to match the
  module text edit dialog.
* Improve the layout of the Pcbnew graphic text dialog.
* Fix some wxPrintf parameter type assertions for 64 bit time_t.
* Fix all (at least I think I got all of them) duplicate menu accelerator
  characters.

+++ eeschema/sch_sheet_path.cpp
@@ -202,7 +202,7 @@ wxString SCH_SHEET_PATH::Path() const
     // it's timestamp changes anyway.
     for( unsigned i = 1; i < m_numSheets; i++ )
     {
-        t.Printf( _( "%8.8lX/" ), m_sheets[i]->GetTimeStamp() );
+        t.Printf( _( "%8.8lX/" ), (long unsigned) m_sheets[i]->GetTimeStamp() );
         s = s + t;
     }
